+++ src/libmach/7obj.c
@@ -83,7 +83,7 @@ static Addr
 addr(Biobuf *bp)
 {
 	Addr a;
-	long off;
+	int32 off;
 
 	a.type = Bgetc(bp);	/* a.type */
 	skip(bp,1);		/* reg */
